+++ crates/wdl-lint/src/util.rs
@@ -8,12 +8,18 @@ use wdl_ast::SyntaxKind;
 /// whitespace.
 pub fn is_inline_comment(token: &Comment) -> bool {
     if let Some(prior) = token.syntax().prev_sibling_or_token() {
-        return prior.kind() != SyntaxKind::Whitespace
-            || !prior
-                .as_token()
-                .expect("should be a token")
-                .text()
-                .contains('\n');
+        let whitespace = prior.kind() == SyntaxKind::Whitespace;
+        if !whitespace {
+            return true;
+        }
+
+        let contains_newline = prior
+            .as_token()
+            .expect("whitespace should be a token")
+            .text()
+            .contains('\n');
+        let first = prior.prev_sibling_or_token().is_none();
+        return !contains_newline && !first;
     }
     false
 }
@@ -64,6 +70,45 @@ mod test {
     use pretty_assertions::assert_eq;
 
     use super::*;
+
+    #[test]
+    fn it_detects_inline() {
+        let (tree, _) = wdl_ast::SyntaxTree::parse(
+            r#"      # not an in-line comment
+
+version 1.2
+
+task foo {  # an in-line comment
+    # not an in-line comment
+}"#,
+        );
+
+        let mut comments = tree
+            .root()
+            .descendants_with_tokens()
+            .filter(|t| t.kind() == SyntaxKind::Comment);
+
+        let first = comments.next().expect("there should be a first comment");
+        let first = Comment::cast(first.as_token().unwrap().clone()).unwrap();
+
+        let is_inline = is_inline_comment(&first);
+
+        assert!(!is_inline);
+
+        let second = comments.next().expect("there should be a second comment");
+        let second = Comment::cast(second.as_token().unwrap().clone()).unwrap();
+
+        let is_inline = is_inline_comment(&second);
+
+        assert!(is_inline);
+
+        let third = comments.next().expect("there should be a third comment");
+        let third = Comment::cast(third.as_token().unwrap().clone()).unwrap();
+
+        let is_inline = is_inline_comment(&third);
+
+        assert!(!is_inline);
+    }
 
     #[test]
     fn test_strip_newline() {